v0.8.4c - scattered card balance adjustments
Overdrive and Cast Off Regrets are nerfed, Seventeen Article Laser and
Amassed Masks are decently buffed. The last of those is probably too much
for now, but it's hard to actually tell players in-game that the average
power on Miko super-scales and costs 1.75.

Really need to go grab some .jsons, soon.