Nerf Grandeur, buff Immortals, adjust Transpose / False Death sound effects
Stacking Grandeur is too reliable, which is why the card had exhaust in the
first place, and with exhaust gone the capacity to stack copies of it for
large stat gains is far too reliable and strong with Artifact support.
Innate was originally intended to help with Miko's focus on set-up, but
the free set-up possibilties just are too much.

Immortal cards getting a mild bump up should help with the Commons pool
being so synergy and combo-focused, as passable sheer quality even without
their scaling effects. Their scaling is slow enough that as a whole it's
hardly affected and people picking them more should help with not seeing
them as a discrete niche.

The two SFX here were too loud- my sound levels over here are a mess, so
I didn't quite check their volumes ahead of time. They've both been
lowered.