Dynamic shifts to Affinity, Yin, Yang
Miko has many keywords and new mechanics, as differentiating her heavily
from many other characters was a deliberate design goal. Feedback, however,
has noted she has multiple characters worth of themes and mechanics, which
pair roughly with rather dense cards. As such, some of her key mechanics
are getting some streamlining.

Affinity in particular is a sticking point for many people. Multiple clauses
on a given tooltip are not easy to absorb, much less two if-then-when pieces,
and the per-turn equalizing of Yin and Yang can further muddle this. The
presence of Hidden Power as a starting card, while meant to be highly thematic
and crucial for starting balance + easing deckbuilding purposes, also rather
annoys newcomers for said complexity also not having any effect for a starting
deck.

As such, Yin and Yang no longer equalize, and instead apply Affinity's effects
at a base 1 by default. This divvies up the burden on Affinity to their
relevant keywords, and it permits moving Hidden Power out of the starting deck
and into the normal card pool. This accompanies buffs to the weaker Affinity
sources, since there's new competition, and nerfs to the Cloaks plus better
Yin / Yang sources, since the starter deck is cleaner and Affinity is more
consistently available.

The following cards have been adjusted in the wake of this change.

 * Red Cloak: 3(5) block -> 2(4) block.
 * Blue Cloak: 3(5) damage -> 2(4) damage.
   [These are still some of the best starting cards of any character,
   and it's tempting to remove their draw in the future. For now, this'll
   probably be fine. Incidentally, Choose Your End is nerfed here too,
   though it's probably still one of the best uncommons around.]

 * Syncretic Surge, Syncretic Plunge: Now uncommon.
   [Even with their prior nerfs, their multiple scalars and synergies
   make them way too strong for common, especially now with Affinity access.
   New Yin and Yang commons will fill in this gap.]

 * Bishamon's Blessing, Guanyin's Blessing:
   Base cards draw 1 card and exhaust when drawn.
   [These should be much less effective being dropped into any deck now, in
   the upgrade-thirsting card layout Miko is left with: more Transpose
   source emphasis, and it makes the upgrade less surprising.
   If this works out, a future card will borrow the mechanic.]

 * Just Rewards: 2(3) Daybreaks ->  2 Daybreaks(+)
   Wicked Punishment: 3 Yin twice (thrice) -> 3(4) Yin twice.
   [Easier Affinity access makes instant triple-sources with benefits
   excessive, and they've got their core value still. It's a shame Just Rewards
   no longer has a hidden Transpose effect built in, but it's still very
   close, and it provides a balance in terms of upgraded card-gen sources.]

...

 * Hidden Power: Now a 1(0)-cost uncommon power.
   Gain 2 Affinity. Draw 1 less card next turn.
   [This is an experiment in offering a cheap if self-debilitating option
    for Affinity, in contrast to the current expensive multipurpose options.
    Whether or not this'll work for the normal card pool remains to be seen,
    but multiple Yin and Yang sources are also nerfed here.]

 * Feast of Dew: Still a 1-cost uncommon skill.
   Draw 2 cards, then add a copy of this card to your Draw Pile (to your hand).
   Exhaust.
   Whenever you play any three Feast of Dew(+) this combat, gain 4 Affinity.
   [This should make the unupgraded version a fair ways more playable than before,
   and make the upgrade less of a sudden surprise compared to the base.]

 * Harmonic Reverence: 3(4) Affinity -> 3(5) Affinity.
   [While there's a number of cards that extremely value the doubling effect,
   it has major playability competition against the lower-cost Hidden Power,
   so it needs a little kick. There's a lot more commons and rares particularly
   wanting for upgrades these days, so this shouldn't be too much.]

I suspect that such Affinity, Yin, and Yang sources might need further nerfs
in light of such increased consistency, when paired with how available they
are when paired with the new five Yin & Yang cards.
For now, we'll see how these changes play out.