v0.9.2: Summary / readme update (Effects flood, Simplified Chinese)
Aside from the few new pieces of content and adjustments, this commit
adds the Chinese translation to the summary (Prince Shoutoku was enamoured
with China, after all) and a pile of allusions / development annotations.

This also updates the .readme a fair bit- previous versions missed adding
much of this.

Whether another version of a few more card / simplifications will come
along, or whether I'll start work on my own (touhou) deckbuilder
roguelike, I can't say. This is probably it for now, barring a bunch of
quick fix updates or a substantial flood of interest.